requirements.txt
pytorch_pretrained_bert==0.6.2
absl-py==0.9.0
backcall==0.2.0
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cymem==2.0.3
decorator==4.4.2
docopt==0.6.2
filelock==3.0.12
future==0.18.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
grpcio==1.30.0
idna==2.10
imageio==2.8.0
importlib-metadata==1.7.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
joblib==0.15.1
Markdown==3.2.2
murmurhash==1.0.2
nltk==3.5
numpy==1.19.0
oauthlib==3.1.0
overrides==3.1.0
packaging==20.4
pandas==1.0.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-crf==0.7.2
pytorch-lightning==0.7.6
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.5.0
sentencepiece==0.1.91
six==1.15.0
sklearn==0.0
spacy==2.3.0
srsly==1.0.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
test-tube==0.7.5
thinc==7.4.1
threadpoolctl==2.1.0
torchtext==0.5.0
tqdm==4.47.0
traitlets==4.3.3
transformers==2.6.0
tokenizers==0.5.2
urllib3==1.25.9
wasabi==0.7.0
wcwidth==0.2.5
Werkzeug==1.0.1
wget==3.2
zenodo-get==1.3.0
zipp==3.1.0
conllu==1.3.1
Unidecode==1.1.1
Flask_Cors==3.0.8
responses==0.10.9
jsonpickle==1.2
boto3==1.10.45
editdistance==0.5.3
gevent==1.4.0
word2number==1.1
matplotlib==3.1.2
ftfy==5.6
numpydoc==0.9.2
botocore==1.13.45
Flask==1.1.1
h5py==2.10.0
Sphinx==2.3.1
parsimonious==0.8.1
thinqpbo==0.1.3
jsonnet==0.14.0
sqlparse==0.3.0
pytest==5.3.2
flaky==3.6.1
docopt==0.6.2
rouge==0.3.2
sphinx_rtd_theme==0.4.3
tensorboardX==1.9
easy_rouge==0.2.2
